Inititial Commit
Created band objects and displayed them in a list box. When I was initializing the objects, I stored each band member name separately in a list. However, this made the ToString( ) to the listbox show up the list's name instead of the member names so I had to restore the list as an array and pass that.